First Working State
Simply this file is included on a webpage and you can start to call it
via CCmenu on global. Sturcture example is below:
CCmenu(Selector, [Array of menu objects]);

Selector: may be a html node or a Jquery selector.

Array of menu objects are predefined as below example:
{item: "some text",
callBack: function(el){
argument el is target object;
things to execute when selected;}
}